% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pairs.R
\name{clean_random_pairs}
\alias{clean_random_pairs}
\title{Clean the random pairs of names to ensure each name only appears once.}
\usage{
clean_random_pairs(rpairs)
}
\arguments{
\item{rpairs}{A two column data frame containing pairs of names as strings on each row.}
}
\value{
A data frame.
}
\description{
Clean the random pairs of names to ensure each name only appears once.
}
