% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_pairs.R
\name{generate_random_pairs}
\alias{generate_random_pairs}
\title{Generate all possible unique random pairs from a set of names, with unwanted pairs removed, if required}
\usage{
generate_random_pairs(names, unwantedpairs = NULL)
}
\arguments{
\item{names}{A vector containing names as strings.}

\item{unwantedpairs}{A two column data frame containing unwanted pairs of names as strings.}
}
\value{
The function returns a two column data frame of random pairs of names as strings.
}
\description{
Generate all possible unique random pairs from a set of names, with unwanted pairs removed, if required
}
