% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_pairs.R
\name{main}
\alias{main}
\title{From a list of names, and unwanted pairs of names if required, generate random pairs,
and save them to a collection of unwanted pairs to be avoided in future.}
\usage{
main(names_file, unwantedpairs_file = NULL)
}
\arguments{
\item{names_file}{Path to a file containing a list of names in a single column.}

\item{unwantedpairs_file}{Path to a file containing a list of unwanted pairs, with
one name per column.}
}
\value{
Writes a csv file containing random pairs of names, and appends to or creates a file
containing unwanted pairs.
}
\description{
From a list of names, and unwanted pairs of names if required, generate random pairs,
and save them to a collection of unwanted pairs to be avoided in future.
}
\examples{
\dontrun{
random_pairs(filename = 'names.csv', unwantedpairs = 'unwantedpairs.csv')
}
}
