% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_pairs.R
\name{random_pairs}
\alias{random_pairs}
\title{Randomly pair up a group of people}
\usage{
random_pairs(filename = NULL, unwantedpairs = NULL)
}
\arguments{
\item{filename}{csv file containing the list of people's names.}

\item{unwantedpairs}{csv file containing pairs of names which you don't want to appear the next time the function is run.}
}
\value{
The function returns nothing, but saves a csv file containing the list of random pairs called \code{Round.csv}, and a csv
file called \code{unwantedpairs.csv} containing pairs which you don't want to appear in future.
}
\description{
Given a list of people's names, pair them up with each other randomly so that each person appears once, and write the
pairs to a file. Also, append these pairs to an additional file, so that they won't be matched
again if the function is rerun.
}
\examples{
\dontrun{
random_pairs(filename = 'names.csv', unwantedpairs = 'unwantedpairs.csv')
}

}
